fix: swap filenames for book-hold canceling and canceled
for some reason they were switched in places.
