+++ libs/lending/domain/src/index.ts
@@ -4,8 +4,8 @@ export * from './lib/book/book-on-hold';
 export * from './lib/book/book-duplicate-hold-found.event';
 export * from './lib/events/book-check-out-failed';
 export * from './lib/events/book-checked-out';
-export * from './lib/events/book-hold-canceled';
 export * from './lib/events/book-hold-canceling-failed';
+export * from './lib/events/book-hold-canceled';
 export * from './lib/events/book-hold-failed';
 export * from './lib/events/book-placed-on-hold';
 export * from './lib/events/book-placed-on-hold-events';
