+++ libs/lending/domain/src/lib/patron.ts
@@ -4,8 +4,8 @@ import { AvailableBook } from './book/available-book';
 import { BookOnHold } from './book/book-on-hold';
 import { BookCheckOutFailed } from './events/book-check-out-failed';
 import { BookCheckedOut } from './events/book-checked-out';
-import { BookHoldCancelingFailed } from './events/book-hold-canceled';
-import { BookHoldCanceled } from './events/book-hold-canceling-failed';
+import { BookHoldCancelingFailed } from './events/book-hold-canceling-failed';
+import { BookHoldCanceled } from './events/book-hold-canceled';
 import { BookHoldFailed } from './events/book-hold-failed';
 import { BookPlacedOnHold } from './events/book-placed-on-hold';
 import { BookPlacedOnHoldEvents } from './events/book-placed-on-hold-events';
